% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_replicates_counts_proportions.R
\name{get_replicate_counts_proportions_Seurat_V5}
\alias{get_replicate_counts_proportions_Seurat_V5}
\title{Collect aggregated data from Seurat Object V4, an internal function in Prop2Count}
\usage{
get_replicate_counts_proportions_Seurat_V5(
  object,
  idents,
  min_cells = 10,
  sep = "__"
)
}
\arguments{
\item{object}{A Seurat object.}

\item{idents}{The colname(s) in the metadata to use as identifiers for aggregation & calculating proportions, what I call replicates here}

\item{min_cells}{A number, the minimum number of barcodes in a unique replicate (see idents) that will be aggregated, fewer than this number will be excluded from the data}

\item{sep}{A character to separate the ident names, if there's more than one metadata column being used. Pick a pretty unique thing that's easy to separate on later}
}
\value{
A list containing the following items:
\itemize{
\item \code{counts}: the total counts per replicate in each gene
\item \code{proportions}: the proportion of cells in the replicate with >0 counts in each gene
\item \code{nCells}: the total cells in the replicate
}
}
\description{
Collect aggregated data from Seurat Object V4, an internal function in Prop2Count
}
\examples{
\dontrun{
   L4_CeNGEN_data <- Data(L4_CeNGEN_data)
   example_list <- get_replicate_counts_proportions_Seurat(L4_CeNGEN_data, c('Cell.type',
                                                           'Experiment'))
}
}
