% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{prevent_infinite}
\alias{prevent_infinite}
\title{Checks for instances where proportions == 1, and moves the value down to prevent infinite counts}
\usage{
prevent_infinite(proportions_matrix, nCell_vector)
}
\arguments{
\item{proportions_matrix}{a genes x replicates matrix of proportions [0,1].}

\item{nCell_vector}{a named vector with integers showing the number of cells in each replicate.}
}
\value{
A matrix of proportions with all proportions less than 1, modified proportions will be returned as nCells-1/nCells
}
\description{
Checks for instances where proportions == 1, and moves the value down to prevent infinite counts
}
