% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{prop2count}
\alias{prop2count}
\title{Converts proportions to a counts-like space}
\usage{
prop2count(proportions_matrix, nCell_vector, round = FALSE)
}
\arguments{
\item{proportions_matrix}{a genes x replicates matrix of proportions from 0,1.}

\item{nCell_vector}{a named vector with integers showing the number of cells in each replicate.}

\item{round}{a boolean value indicating whether or not to round the output down to an integer value}
}
\value{
a matrix of values from 0,infinity to be used as counts for downstream applications
}
\description{
Converts proportions to a counts-like space
}
