% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop2Count.R
\name{show,Prop2Count-method}
\alias{show,Prop2Count-method}
\title{Show method for Prop2Count objects}
\usage{
\S4method{show}{Prop2Count}(object)
}
\arguments{
\item{object}{A Prop2Count object}
}
\description{
Show method for Prop2Count objects
}
