% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop2Count.R
\name{create_Prop2Count}
\alias{create_Prop2Count}
\title{Collect aggregated data from Seurat Object, an internal function in Prop2Count}
\usage{
create_Prop2Count(
  object,
  idents,
  sep = "__",
  min_cells = 10,
  round = FALSE,
  keep_counts = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{idents}{The colname(s) in the metadata to use as identifiers for aggregation & calculating proportions, what I call replicates here}

\item{sep}{A character to separate the ident names, if there's more than one metadata column being used. Pick a pretty unique thing that's easy to separate on later}

\item{min_cells}{A number, the minimum number of barcodes in a unique replicate (see idents) that will be aggregated, fewer than this number will be excluded from the data}

\item{round}{A boolean value, indicating whether or not to round the transformed proportions output down to an integer. Default = FALSE}

\item{keep_counts}{A boolean value, indicating whether to keep the oriignal aggregated counts matrix. Default = TRUE}
}
\value{
a Prop2Count object with transformed counts
}
\description{
Collect aggregated data from Seurat Object, an internal function in Prop2Count
}
